+++ NEWS.md
@@ -2,6 +2,7 @@
 ## New features
 
 * qrandommaxint (true random signed integers)
+* qUUID (true random Universally Unique IDs (UUID))
 
 # qrandom 1.0
 ## New features